Address test-support review findings (#77)
Clarify that `proptest` is a `tei-py` dev-dependency, remove
fixed-count bootstrap tests covered by the property tests, and split the
`test_support` unit tests into focused submodules so each source file
stays under the project line limit.

Keep the `msgspec_available` API unchanged because redesigning that
public query helper is outside this branch's property-test scope.
