+++ tei-py/src/test_support.rs
@@ -177,412 +177,5 @@ pub fn msgspec_available() -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    //! Unit tests for Python-side test support helpers.
-
-    use super::*;
-    use proptest::prelude::*;
-    use pyo3::{pyclass, pymethods};
-    use rstest::rstest;
-    use std::{
-        ffi::CString,
-        sync::{
-            Arc,
-            atomic::{AtomicUsize, Ordering},
-        },
-        thread,
-    };
-
-    struct RunAndKwargs<'py> {
-        run: Bound<'py, PyAny>,
-        kwargs: Bound<'py, PyDict>,
-        globals: Bound<'py, PyDict>,
-    }
-
-    #[derive(Clone, Copy)]
-    enum RunWithKwargsArgShape {
-        Unit,
-        NestedPyTuple,
-        DirectPyTuple,
-    }
-
-    fn setup_run_and_kwargs(py: Python<'_>) -> RunAndKwargs<'_> {
-        let globals = PyDict::new(py);
-        let patch = CString::new(
-            "import subprocess\n\
-             _original_run = subprocess.run\n\
-             _calls = []\n\
-             subprocess.run = lambda *a, **kw: _calls.append((a, kw))\n",
-        )
-        .expect("CString build");
-        py.run(patch.as_c_str(), Some(&globals), None)
-            .expect("monkeypatch subprocess.run");
-        let subprocess = py.import("subprocess").expect("import subprocess");
-        let run = subprocess.getattr("run").expect("get subprocess.run");
-        let kwargs = PyDict::new(py);
-
-        RunAndKwargs {
-            run,
-            kwargs,
-            globals,
-        }
-    }
-
-    fn restore_subprocess_run(py: Python<'_>, globals: &Bound<'_, PyDict>) {
-        // Restore `subprocess.run` and remove the bootstrap `meta_path` blocker
-        // if one was installed. The blocker removal is a no-op for callers that
-        // never installed it (e.g. the `run_with_kwargs` tests).
-        let restore = CString::new(
-            r"
-import subprocess
-import sys
-
-subprocess.run = _original_run
-_calls = []
-
-try:
-    sys.meta_path.remove(_bootstrap_msgspec_blocker)
-except (ValueError, NameError):
-    pass
-",
-        )
-        .expect("CString build");
-        py.run(restore.as_c_str(), Some(globals), None).ok();
-    }
-
-    struct SubprocessRestoreGuard<'py> {
-        py: Python<'py>,
-        globals: Bound<'py, PyDict>,
-    }
-
-    impl Drop for SubprocessRestoreGuard<'_> {
-        fn drop(&mut self) {
-            restore_subprocess_run(self.py, &self.globals);
-        }
-    }
-
-    #[pyclass]
-    struct BootstrapRunCounter {
-        count: Arc<AtomicUsize>,
-    }
-
-    #[pymethods]
-    impl BootstrapRunCounter {
-        fn __call__(&self, _args: &Bound<'_, PyTuple>, _kwargs: Option<&Bound<'_, PyDict>>) {
-            self.count.fetch_add(1, Ordering::SeqCst);
-        }
-    }
-
-    fn setup_bootstrap_run_counter(py: Python<'_>, count: Arc<AtomicUsize>) -> pyo3::Py<PyDict> {
-        let globals = PyDict::new(py);
-        globals
-            .set_item(
-                "_counter",
-                pyo3::Py::new(py, BootstrapRunCounter { count }).expect("build run counter"),
-            )
-            .expect("install run counter");
-        // Force the bootstrap to actually run, then count and satisfy it:
-        //   * a `meta_path` blocker makes `import msgspec` fail, so
-        //     `ensure_msgspec_installed` cannot short-circuit and must enter the
-        //     `Once`-guarded installer (even if msgspec is installed on disk);
-        //   * the mocked `subprocess.run` counts each call and registers an
-        //     importable stub `msgspec` module, so the installer's final
-        //     `import msgspec` resolves without network access.
-        // The counter therefore proves the installer path executed, and the
-        // assertions fail if the bootstrap is ever skipped. Each test runs in
-        // its own process under nextest, so the process-wide `Once` is fresh and
-        // no cross-test reset seam or mutex is required.
-        let patch = CString::new(
-            r#"
-import subprocess
-import sys
-import types
-
-
-class _BlockMsgspecBootstrap:
-    def find_spec(self, fullname, path=None, target=None):
-        if fullname == "msgspec" or fullname.startswith("msgspec."):
-            raise ModuleNotFoundError("msgspec blocked for bootstrap test", name="msgspec")
-        return None
-
-
-_bootstrap_msgspec_blocker = _BlockMsgspecBootstrap()
-sys.meta_path.insert(0, _bootstrap_msgspec_blocker)
-
-_original_run = subprocess.run
-
-
-def _mock_run(*args, **kwargs):
-    _counter(args, kwargs)
-    sys.modules.setdefault("msgspec", types.ModuleType("msgspec"))
-
-
-subprocess.run = _mock_run
-"#,
-        )
-        .expect("CString build");
-        py.run(patch.as_c_str(), Some(&globals), None)
-            .expect("monkeypatch subprocess.run");
-
-        globals.unbind()
-    }
-
-    fn remove_msgspec_from_modules(py: Python<'_>) {
-        let remove =
-            CString::new("import sys\nsys.modules.pop('msgspec', None)\n").expect("CString build");
-        py.run(remove.as_c_str(), None, None)
-            .expect("remove msgspec from sys.modules");
-    }
-
-    fn recorded_call_count(globals: &Bound<'_, PyDict>) -> usize {
-        globals
-            .get_item("_calls")
-            .expect("read recorded subprocess.run calls")
-            .len()
-            .expect("count recorded subprocess.run calls")
-    }
-
-    fn recorded_args<'py>(globals: &Bound<'py, PyDict>) -> Bound<'py, PyAny> {
-        let expr = CString::new("_calls[0][0]").expect("CString build");
-        globals
-            .py()
-            .eval(expr.as_c_str(), Some(globals), None)
-            .expect("read recorded subprocess.run positional arguments")
-    }
-
-    #[rstest]
-    #[case::unit_tuple(RunWithKwargsArgShape::Unit)]
-    #[case::one_tuple_of_pytuple(RunWithKwargsArgShape::NestedPyTuple)]
-    #[case::bound_pytuple(RunWithKwargsArgShape::DirectPyTuple)]
-    fn run_with_kwargs_accepts_supported_arg_shapes(#[case] arg_shape: RunWithKwargsArgShape) {
-        Python::attach(|py| {
-            let RunAndKwargs {
-                run,
-                kwargs,
-                globals,
-            } = setup_run_and_kwargs(py);
-            let _restore_guard = SubprocessRestoreGuard {
-                py,
-                globals: globals.clone(),
-            };
-
-            match arg_shape {
-                RunWithKwargsArgShape::Unit => run_with_kwargs(&run, (), &kwargs),
-                RunWithKwargsArgShape::NestedPyTuple => {
-                    let args_tuple = PyTuple::new(py, ["true"]).expect("build argument tuple");
-
-                    run_with_kwargs(&run, (args_tuple,), &kwargs);
-                }
-                RunWithKwargsArgShape::DirectPyTuple => {
-                    let args_tuple = PyTuple::new(py, [["true"]]).expect("build subprocess args");
-
-                    run_with_kwargs(&run, args_tuple, &kwargs);
-                }
-            }
-
-            let call_count = recorded_call_count(&globals);
-
-            assert_eq!(call_count, 1);
-
-            let args = recorded_args(&globals);
-            match arg_shape {
-                RunWithKwargsArgShape::Unit => {
-                    assert_eq!(args.len().expect("count positional arguments"), 0);
-                }
-                RunWithKwargsArgShape::NestedPyTuple => {
-                    let first_arg = args.get_item(0).expect("read first positional argument");
-                    assert_eq!(
-                        first_arg
-                            .extract::<(String,)>()
-                            .expect("extract nested tuple argument"),
-                        ("true".to_owned(),)
-                    );
-                }
-                RunWithKwargsArgShape::DirectPyTuple => {
-                    assert_eq!(args.len().expect("count positional arguments"), 1);
-                    let first_arg = args.get_item(0).expect("read first positional argument");
-                    assert_eq!(
-                        first_arg
-                            .extract::<Vec<String>>()
-                            .expect("extract direct list argument"),
-                        vec!["true".to_owned()]
-                    );
-                }
-            }
-        });
-    }
-
-    #[test]
-    fn ensure_msgspec_installed_invokes_subprocess_at_most_once_across_repeated_calls() {
-        let run_count = Arc::new(AtomicUsize::new(0));
-
-        let globals = Python::attach(|py| {
-            let g = setup_bootstrap_run_counter(py, Arc::clone(&run_count));
-            remove_msgspec_from_modules(py);
-            g
-        });
-
-        assert!(Python::attach(ensure_msgspec_installed).is_ok());
-        assert!(Python::attach(ensure_msgspec_installed).is_ok());
-
-        assert!(
-            Python::attach(|py| {
-                restore_subprocess_run(py, globals.bind(py));
-                ensure_msgspec_installed(py)
-            })
-            .is_ok()
-        );
-
-        // The blocker forces the installer to run exactly once: subprocess.run
-        // is invoked for ensurepip and for the msgspec install. Asserting the
-        // exact count proves the bootstrap path executed (it would be 0 if
-        // skipped) and that the `Once` guard prevented a second bootstrap across
-        // both repeated calls.
-        assert_eq!(run_count.load(Ordering::SeqCst), 2);
-    }
-
-    proptest! {
-        #![proptest_config(ProptestConfig {
-            // `MSGSPEC_INIT` is process-wide, so each generated case needs its
-            // own nextest process to exercise the forced bootstrap path.
-            cases: 1,
-            ..ProptestConfig::default()
-        })]
-
-        #[test]
-        fn idempotency_holds_over_arbitrary_repetitions(repetitions in 1..=50u8) {
-            let run_count = Arc::new(AtomicUsize::new(0));
-
-            let globals = Python::attach(|py| {
-                let g = setup_bootstrap_run_counter(py, Arc::clone(&run_count));
-                remove_msgspec_from_modules(py);
-                g
-            });
-
-            for _ in 0..usize::from(repetitions) {
-                let result = Python::attach(ensure_msgspec_installed);
-                prop_assert!(result.is_ok());
-            }
-
-            let restored_after_bootstrap = Python::attach(|py| {
-                restore_subprocess_run(py, globals.bind(py));
-                ensure_msgspec_installed(py)
-            })
-            .is_ok();
-            prop_assert!(restored_after_bootstrap);
-
-            prop_assert_eq!(run_count.load(Ordering::SeqCst), 2);
-        }
-
-        #[test]
-        fn no_panics_under_variable_thread_count(thread_count in 2..=32u8) {
-            let run_count = Arc::new(AtomicUsize::new(0));
-            let globals = Python::attach(|py| {
-                let globals = setup_bootstrap_run_counter(py, Arc::clone(&run_count));
-                remove_msgspec_from_modules(py);
-
-                globals
-            });
-
-            let handles: Vec<_> = (0..usize::from(thread_count))
-                .map(|_| thread::spawn(move || Python::attach(ensure_msgspec_installed)))
-                .collect();
-
-            for handle in handles {
-                let result = handle.join().expect("bootstrap thread panicked");
-                prop_assert!(result.is_ok());
-            }
-
-            let restored_after_bootstrap = Python::attach(|py| {
-                restore_subprocess_run(py, globals.bind(py));
-                ensure_msgspec_installed(py)
-            })
-            .is_ok();
-            prop_assert!(restored_after_bootstrap);
-
-            prop_assert_eq!(run_count.load(Ordering::SeqCst), 2);
-        }
-    }
-
-    #[test]
-    fn msgspec_available_reports_true_only_when_msgspec_is_importable() {
-        // Call the function under test first; it may bootstrap msgspec as a
-        // side-effect, so the importability check must come *after* the call.
-        let reported_available = msgspec_available();
-        let importable_after_check = Python::attach(|py| py.import("msgspec").is_ok());
-
-        assert_eq!(reported_available, importable_after_check);
-    }
-
-    #[test]
-    fn ensure_msgspec_installed_is_safe_under_concurrent_access() {
-        let run_count = Arc::new(AtomicUsize::new(0));
-        let globals = Python::attach(|py| {
-            let globals = setup_bootstrap_run_counter(py, Arc::clone(&run_count));
-            remove_msgspec_from_modules(py);
-
-            globals
-        });
-
-        let handles: Vec<_> = (0..8)
-            .map(|_| thread::spawn(move || Python::attach(ensure_msgspec_installed)))
-            .collect();
-
-        for handle in handles {
-            assert!(handle.join().expect("bootstrap thread panicked").is_ok());
-        }
-
-        assert!(
-            Python::attach(|py| {
-                restore_subprocess_run(py, globals.bind(py));
-                ensure_msgspec_installed(py)
-            })
-            .is_ok()
-        );
-
-        // The blocker forces the installer to run, and the `Once` guard fires
-        // exactly once across all threads, making exactly two subprocess.run
-        // calls (ensurepip + msgspec install). The exact count proves the
-        // bootstrap path executed rather than being skipped.
-        assert_eq!(run_count.load(Ordering::SeqCst), 2);
-    }
-
-    #[rstest]
-    #[case::none_means_absent("None", false)]
-    #[case::path_means_present("'/usr/bin/uv'", true)]
-    fn has_uv_reflects_which_return_value(#[case] which_return_expr: &str, #[case] expected: bool) {
-        Python::attach(|py| {
-            let globals = PyDict::new(py);
-            let patch = CString::new(format!(
-                "import shutil\n\
-                 orig = shutil.which\n\
-                 shutil.which = lambda name: {which_return_expr}\n"
-            ))
-            .expect("CString build");
-            py.run(patch.as_c_str(), Some(&globals), None)
-                .expect("monkeypatch shutil.which");
-            let _restore_guard = ShutilRestoreGuard {
-                py,
-                globals: globals.clone(),
-            };
-
-            assert_eq!(has_uv(py), expected);
-        });
-    }
-
-    fn restore_shutil_which(py: Python<'_>, globals: &Bound<'_, PyDict>) {
-        let restore = CString::new("import shutil\nshutil.which = orig\n").expect("CString build");
-        py.run(restore.as_c_str(), Some(globals), None).ok();
-    }
-
-    struct ShutilRestoreGuard<'py> {
-        py: Python<'py>,
-        globals: Bound<'py, PyDict>,
-    }
-
-    impl Drop for ShutilRestoreGuard<'_> {
-        fn drop(&mut self) {
-            restore_shutil_which(self.py, &self.globals);
-        }
-    }
-}
+#[path = "test_support_tests/mod.rs"]
+mod tests;
